pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "carms-explorer"
version = "0.1.0"
description = "AI-powered CaRMS residency program explorer"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    "pydantic-settings>=2.0",
    "alembic>=1.13",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.0",
    "httpx>=0.27",
    "sentence-transformers>=3.0",
]
dagster = [
    "dagster>=1.9",
    "dagster-webserver>=1.9",
    "dagster-postgres>=0.25",
    "openpyxl>=3.1",
    "pandas>=2.2",
    "langchain-text-splitters>=0.3",
    "sentence-transformers>=3.0",
]
rag = [
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-community>=0.3",
]
agent = [
    "claude-agent-sdk>=0.1",
    "anthropic>=0.42",
]
dashboard = [
    "streamlit>=1.40",
    "plotly>=5.24",
    "pandas>=2.2",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
    "pymdown-extensions>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.27",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/carms"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/carms"]
omit = [
    "src/carms/agent/*",
    "src/carms/dashboard/*",
    "src/carms/etl/*",
    "src/carms/search/rag.py",
]

[tool.coverage.report]
fail_under = 70
